wait for registry to catch up
Sometimes, the registry isn't quite up to date with the changes feed, so
when follow-registry makes a request to the registry to get the latest
metadata, it's one version too old. This change checks the _rev on the
metadata against the rev provided in the changes feed. It's it's too
old, it waits a minute and tries again.