Use normalize-registry-metadata
npm's skimdb and registry are not actually consistent. skimdb is
the 'raw' data, which, among other things, includes references to
versions that were published before npm began enforcing proper
semver syntax (`1.0.0beta` vs 1.0.0-beta`). The public registry
contains a sanitized version of the skimdb, with the documents
piped through the published `normalize-registry-metadata` library.

This patch adds this same pipe to the follow-registry, ensuring
that documents returned from the registry are in a consistent state
before being handed off to any downstream consumers. Without it,
a downstream consumer is likely to have unexpected behavior, as
they believe the versions match the registry, when they actually
do not.

Note that downstream mirrors, depending on their implementation,
may have to rebuild their indexes. Using the normalize library
does not appear to impact the name of the tarballs, thus this
reconstruction would be restricted to the index.json files only.